+++ src/asp/Tools/stereo_pprc.cc
@@ -588,36 +588,47 @@ void stereo_preprocessing(bool adjust_left_image_size, ASPGlobalOptions& opt) {
     write_vector(right_stats_file, right_stats2);
   }
 
+  // When alignment method is none or epipolar, no ip were created so far, so
+  // produce them now.
+  if (stereo_settings().alignment_method == "none" ||
+      stereo_settings().alignment_method == "epipolar")
+    asp::compute_ip_LR(opt.out_prefix);
+    
 } // End function stereo_preprocessing
 
 // Find the median angle in degrees at which rays emanating from
 // matching points meet
 void estimate_convergence_angle(ASPGlobalOptions const& opt) {
 
-  // For alignment method none the ip match file does not exist yet,
-  // and it is done later only for subsampled images. For alignment
-  // method epipolar need to undo the alignment. Hence don't estimate
-  // the convergence angle for now as that would require major
-  // reshuffling. 
-  bool will_do = (stereo_settings().alignment_method == "homography" ||
-                  stereo_settings().alignment_method == "affineepipolar" ||
-                  stereo_settings().alignment_method == "local_epipolar");
+  if (stereo_settings().correlator_mode)
+    return; // No camera can be assumed, hence no convergence angle.
 
-  if (!will_do)
-    return;
+  // When having matches between L and R, need to do things a bit differently.
+  bool have_aligned_matches = (stereo_settings().alignment_method == "none" ||
+                               stereo_settings().alignment_method == "epipolar");
 
-  std::string match_filename
-    = opt.session->stereo_match_filename(opt.session->left_cropped_image(),
-                                         opt.session->right_cropped_image(),
-                                         opt.out_prefix);
-  
+  std::string match_filename;
+  if (have_aligned_matches)
+    match_filename = vw::ip::match_filename(opt.out_prefix, "L.tif", "R.tif");
+  else 
+    match_filename = opt.session->stereo_match_filename(opt.session->left_cropped_image(),
+                                                        opt.session->right_cropped_image(),
+                                                        opt.out_prefix);
   // The interest points must exist by now
   if (!fs::exists(match_filename))
-    vw_throw(ArgumentErr() << "Missing IP file: " << match_filename);
+    vw_throw(ArgumentErr() << "Missing IP matches file: " << match_filename);
   
   std::vector<ip::InterestPoint> left_ip, right_ip;
-  // vw_out() << "Reading binary match file: " << match_filename << std::endl;
   ip::read_binary_match_file(match_filename, left_ip, right_ip);
+
+  if (have_aligned_matches) {
+    // Unalign the interest point matches
+    std::vector<vw::ip::InterestPoint> unaligned_left_ip, unaligned_right_ip;
+    asp::unalign_ip(opt.session->tx_left(), opt.session->tx_right(),
+                    left_ip, right_ip, unaligned_left_ip, unaligned_right_ip);
+    left_ip  = unaligned_left_ip;
+    right_ip = unaligned_right_ip;
+  }
 
   std::vector<double> sorted_angles;
   boost::shared_ptr<camera::CameraModel> left_cam, right_cam;
@@ -667,8 +678,8 @@ int main(int argc, char* argv[]) {
                                    !stereo_settings().part_of_multiview_run);
 
     stereo_preprocessing(adjust_left_image_size, opt);
+    
     // TODO(oalexan1): Must test with crop win!
-
     estimate_convergence_angle(opt);
     
     vw_out() << "\n[ " << current_posix_time_string() << " ] : PREPROCESSING FINISHED \n";